Add depth to info_dict, depth_nngrid.py and two_channel_depth_nngrid.py
depth_nngrid.py has a single additional grid channel where
body/joint's grid cell has value 0.0 if it's a front part and 1.0 if it's a
"back" part. two_channel_depth_nngrid.py has an alternative with two channels,
where the first channel is 1.0 for the front parts and the second
channel is 1.0 for the back parts.