Fix bug with action grid given by model being different size than state
The edge length of the grid depends on the output of the model rather
than the size of the input grid, so update the action grid to draw with
the edge length of the model's outputted action grid instead of the edge
length of the state grid. Also update visualize's 'load' argument.