Add directory of models, plotting functionality for directory_evaluate
Now directory_evaluate can be run on a directory of models to evaluate
as well as a directory of JSONs, and will plot the all episode reward
means vs model number. Supports evaluating a range of model numbers as
well as all models in a directory.

Note: model checkpoints don't start at 0, so the graph won't start at 0
either