Fix bug where action decoder goes out of bounds
Due to convolution filter size, padding, and stride, discovered that
model outputted action is not always the same dimensions as the input,
which causes an out of bounds error in a very rare case (overlooked
until implementation of true random env), so added a hack (clamping to
size of action) to address this without changing model architectures.
Also, change name of default params file to go along with 43e3138.